+++ src/Resources/PaymentResource.php
@@ -20,7 +20,7 @@ class PaymentResource extends Resource
      * $link = Plorea::payments()
      *     ->link('FIN-2026-00123', 'Faktura FIN-2026-00123', Amount::nok(450000), 'https://app.test/paid')
      *     ->payerEmail('kunde@eksempel.no')
-     *     ->merchant(orgNr: '912650774', name: 'Techify AS', email: 'post@techify.no')
+     *     ->merchant(orgNr: '999999999', name: 'Techify AS', email: 'post@techify.no')
      *     ->create();
      * ```
      */